+++ gateway/graphql/market_resolvers.go
@@ -179,6 +179,10 @@ func (r *myMarketResolver) DecimalPlaces(ctx context.Context, obj *types.Market)
 	return int(obj.DecimalPlaces), nil
 }
 
+func (r *myMarketResolver) PositionDecimalPlaces(ctx context.Context, obj *types.Market) (int, error) {
+	return int(obj.PositionDecimalPlaces), nil
+}
+
 func (r *myMarketResolver) Name(ctx context.Context, obj *types.Market) (string, error) {
 	return obj.TradableInstrument.Instrument.Name, nil
 }